Add files via upload
This script is a simple python calculator that takes an input as to whether you want to Add, Subtract, multiply, or divide two numbers. Once you select your operation it then inputs for you to enter a first number or your x variable and then it prompts you to enter a second number or your Y variable and then the program completes the script and gives you your output. I plan to take the knowledge gained from this script and expand it to be able to run for different financial equations such as a car payment calculator.